+++ lib/src/colors.dart
@@ -4,5 +4,5 @@ final List<String> colorPalette = [
   '#4169E1', //  (Royal Blue)
   '#FFD700', //  (Gold)
   '#8A2BE2', //  (Blue Violet)
-  '#4682B4' //  (Steel Blue)eal
+  '#4682B4' //  (Steel Blue)
 ];